pyproject.toml
[project]
name = "ai-dj"
version = "0.1.5"
description = "AI agent for music curration"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.97.0",
    "elevenlabs>=2.45.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.0",
    "rich>=13.0.0",
    "python-dotenv>=1.2.2",
    "spotipy>=2.26.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_dj", "src/server", "src/client", "src/tui"]

[tool.uv]
package = true

[project.scripts]
ai-dj-server = "server.main:main"
ai-dj = "tui.main:main"
